requirements.txt
asgiref==3.5.2
Django==4.1
django-tinymce==3.4.0
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==6.2.0
Pillow==9.2.0